Add many helpful scalacOptions and fix the resulting compile errors they exposed
Also keep -Xlog-implicits around, in case it's ever needed for
debugging craziness.

+++ src/test/scala/fpInScala/exercises/chapter6/Exercise1Test.scala
@@ -5,7 +5,7 @@ import org.scalacheck.Properties
 import org.scalacheck.Prop.forAll
 
 class Exercise1Test extends Properties("RNG") {
-  property("RNG.nextInt produces a value between 0 and Int.MaxValue, inclusive") = forAll { (n: Int) =>
+  property("RNG.nextInt produces a value between 0 and Int.MaxValue, inclusive") = forAll { (n: Long) =>
     val rng = SimpleRNG(n)
     val (nextRand, _) = Exercise1.nonNegativeInt(rng)
     nextRand >= 0 && nextRand <= Int.MaxValue
