+++ src/test/scala/fpInScala/exercises/chapter6/Exercise2Test.scala
@@ -5,7 +5,7 @@ import org.scalacheck.Properties
 import org.scalacheck.Prop.forAll
 
 class Exercise2Test extends Properties("Exercise2.double") {
-  property("Generates a Double between 0 and 1, not including 1") = forAll { (n: Int) =>
+  property("Generates a Double between 0 and 1, not including 1") = forAll { (n: Long) =>
     val rng = SimpleRNG(n)
     val (nextRand, _) = Exercise2.double(rng)
     nextRand >= 0.0 && nextRand < 1.0
