+++ src/test/scala/fpInScala/exercises/chapter6/Exercise4Test.scala
@@ -5,7 +5,7 @@ import org.scalacheck.Properties
 import org.scalacheck.Prop.forAll
 
 class Exercise4Test extends Properties("Exercise4.ints") {
-  property("Produces a list of random integers, of the count specified") = forAll { (n: Int, c: Int) =>
+  property("Produces a list of random integers, of the count specified") = forAll { (n: Long, c: Int) =>
     val rng = SimpleRNG(n)
     val count = Math.abs(c - 1) % 500 // Keep the lists short-ish
