+++ src/test/scala/fpInScala/exercises/chapter6/Exercise5Test.scala
@@ -5,7 +5,7 @@ import org.scalacheck.Properties
 import org.scalacheck.Prop.forAll
 
 class Exercise5Test extends Properties("Exercise5.double") {
-  property("Generates a Double between 0 and 1, not including 1, in a more elegant way") = forAll { (n: Int) =>
+  property("Generates a Double between 0 and 1, not including 1, in a more elegant way") = forAll { (n: Long) =>
     val rng = SimpleRNG(n)
     val (nextRand, _) = Exercise5.double(rng)
     nextRand >= 0.0 && nextRand < 1.0
