+++ src/test/scala/fpInScala/exercises/chapter6/Exercise8Test.scala
@@ -5,7 +5,7 @@ import org.scalacheck.Properties
 import org.scalacheck.Prop.forAll
 
 class Exercise8Test extends Properties("Exercise8.nonNegativeIntLessThan") {
-  property("Produces a non-negative integer, less than some other non-negative integer") = forAll { (n: Int) =>
+  property("Produces a non-negative integer, less than some other non-negative integer") = forAll { (n: Long) =>
     val rng = SimpleRNG(n)
     val (sentinel, nextRNG) = Exercise1.nonNegativeInt(rng)
