+++ src/test/scala/fpInScala/exercises/chapter6/Exercise9Test.scala
@@ -5,13 +5,13 @@ import org.scalacheck.Properties
 import org.scalacheck.Prop.forAll
 
 class Exercise9Test extends Properties("Exercise9.map and map2 in terms of flatMap") {
-  property("Exercise9.double generates a Double between 0 and 1, not including 1, in a still more elegant way") = forAll { (n: Int) =>
+  property("Exercise9.double generates a Double between 0 and 1, not including 1, in a still more elegant way") = forAll { (n: Long) =>
     val rng = SimpleRNG(n)
     val (nextRand, _) = Exercise9.double(rng)
     nextRand >= 0.0 && nextRand < 1.0
   }
 
-  property("sequence(List(unit(1), unit(2), unit(3)))(rng)._1 should return List(1, 2, 3)") = forAll { (n: Int) =>
+  property("sequence(List(unit(1), unit(2), unit(3)))(rng)._1 should return List(1, 2, 3)") = forAll { (n: Long) =>
     import Exercise9._
 
     val r = SimpleRNG(n)
@@ -19,7 +19,7 @@ class Exercise9Test extends Properties("Exercise9.map and map2 in terms of flatM
   }
 
   property("Produces a list of random integers, of the count specified, implemented in terms of Exercise7.sequence") =
-    forAll { (n: Int, c: Int) =>
+    forAll { (n: Long, c: Int) =>
       val rng = SimpleRNG(n)
       val count = Math.abs(c - 1) % 500 // Keep the lists short-ish
